requirements.txt
annotated-types==0.7.0
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.135.3
passlib[bcrypt]==1.7.4
psycopg2-binary
pydantic==2.13.0
python-jose==3.5.0
python-multipart==0.0.26
SQLAlchemy==2.0.49
uvicorn==0.44.0
bcrypt==4.0.1
email-validator==2.3.0
fastapi==0.135.3
gunicorn
passlib[bcrypt]==1.7.4
psycopg2-binary
pydantic==2.13.0
python-jose==3.5.0
python-multipart==0.0.26
SQLAlchemy==2.0.49
uvicorn==0.44.0
Pillow==12.2.0
